gemini_voice_chatbot/requirements.txt
streamlit
speechrecognition
pyttsx3
google-generativeai
pyaudio
